Moved dispatch inside the main upring object
